Extract the API layer into a shared @lfc/api package
Move the entire lfc.se API layer out of the mobile app into a
transport-agnostic package consumed by both the app and the notifications
backend. Renames @lfc/shared -> @lfc/api and folds in the request layer,
parsers, domain types and errors that previously lived in the app's api.ts.

Auth boundary: the session cookie transport is the only platform-specific
part, so createClient() delegates it. Mobile keeps its native cookie glue and
passes credentials:'include'; a server can instead supply getSessionToken and
the client sends the session as a Cookie header. Public endpoints (news,
fixtures, standings, comments) need no auth; login/logout/member/likeComment
are marked as authenticated.

- packages/api: types, errors, parsers, createClient, plus the existing
  lightweight fetchLatestPosts used by the backend.
- apps/mobile/src/api.ts: reduced to a thin native-cookie adapter that creates
  the client and re-exports its methods, so every ~/api import is unchanged.
- apps/notifications: repointed to @lfc/api.

typecheck (all three packages), mobile eslint, formatting and an offline
end-to-end backend smoke test all pass.
