+++ packages/api/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@lfc/shared",
+  "name": "@lfc/api",
   "version": "1.0.0",
   "private": true,
   "type": "module",
