--- packages/shared/src/index.ts
@@ -1,6 +1,0 @@
-export type { NewsPost } from './types'
-export {
-  DEFAULT_API_URL,
-  fetchLatestPosts,
-  type FetchLatestPostsOptions,
-} from './news'
